Adds deprecated note to `working-with-rules-deprecated` page.
I wound up on the [working-with-rules-deprecated](https://github.com/eslint/eslint/blob/master/docs/developer-guide/working-with-rules-deprecated.md) page via a google search or something, but I didn't notice that the page name had `deprecated` in it. As I was reading, I was like, this doesn't seem right... Took me a good 15 minutes to realize I was reading the old version. I figured adding a note to the top, similar to the note at the top of the [working-with-rules](https://github.com/eslint/eslint/blob/master/docs/developer-guide/working-with-rules.md) page might be helpful in case anyone else winds up here like I did.